pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
python-dotenv = "^0.14.0"
pytest = "^7.4.0"
gunicorn = "^22.0.0"
pymongo = "^4.6.2"
mongomock = "^4.1.2"
safety = "^3.1.0"
loggly-python-handler = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
